Move all json related stuff to JsonHelpers
